+++ lib/ftpd.rb
@@ -7,10 +7,14 @@ require 'socket'
 require 'tmpdir'
 
 module Ftpd
+  module ListFormat
+    autoload :Ls,                      'ftpd/list_format/ls'
+  end
   autoload :CommandSequenceChecker,    'ftpd/command_sequence_checker'
   autoload :DiskFileSystem,            'ftpd/disk_file_system'
   autoload :Error,                     'ftpd/error'
   autoload :ExceptionTranslator,       'ftpd/exception_translator'
+  autoload :FileInfo,                  'ftpd/file_info'
   autoload :FileSystemErrorTranslator, 'ftpd/file_system_error_translator'
   autoload :FileSystemMethodMissing,   'ftpd/file_system_method_missing'
   autoload :FtpServer,                 'ftpd/ftp_server'